% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen1.R
\name{eigen1}
\alias{eigen1}
\alias{eigen1_power_iteration}
\title{Compute leading eigenvalue}
\usage{
eigen1(m, ..., method = "power_iteration")

eigen1_power_iteration(
  m,
  max_iterations = 100,
  tolerance = sqrt(.Machine$double.eps),
  initial = NULL
)
}
\arguments{
\item{m}{A matrix or 3d array}

\item{...}{Additional arguments passed to the underlying method}

\item{method}{Select the method to use. Currently only
"power_iteration" is supported.}

\item{max_iterations}{Integer, giving the number of iterations before
giving up}

\item{tolerance}{Number, giving the required tolerance}

\item{initial}{An optional initial guess at the eigenvector. If
omitted we use a random vector}
}
\value{
A scalar real (if \code{m} is a matrix) or a vector with length
\code{dim(m)[3]} if \code{m} was a 3d array
}
\description{
Compute the leading eigenvalue for a square matrix
}
\examples{
m <- diag(runif(10))
eigen1::eigen1(m)
max(eigen(m)$values)
}
